requirements.txt
python3
python3-pyglet
